Improve CBOR
